requirements.txt
click
rich
GitPython
requests
